+++ packages/core/checkbox/src/__tests__/unit/index.js
@@ -4,21 +4,15 @@ import React from 'react';
 import { mount } from 'enzyme';
 import CheckboxIcon from '@atlaskit/icon/glyph/checkbox';
 import CheckboxIndeterminateIcon from '@atlaskit/icon/glyph/checkbox-indeterminate';
-
-import Checkbox, { CheckboxGroup } from '../../';
-import CheckboxStatelessWithAnalytics, {
-  CheckboxStatelessWithoutAnalytics as CheckboxStateless,
-} from '../../CheckboxStateless';
-import { HiddenCheckbox } from '../../styled/Checkbox';
+import Checkbox from '../../Checkbox';
 import { name } from '../../../package.json';
 
 describe(name, () => {
-  describe('<CheckboxStateless />', () => {
-    const mountStatelessCheckbox = (overridingProps: any) =>
+  describe('<Checkbox />', () => {
+    const mountCheckbox = (overridingProps: any) =>
       mount(
-        <CheckboxStateless
+        <Checkbox
           label=""
-          isChecked
           onChange={() => {}}
           name="stub"
           value="stub value"
@@ -27,91 +21,73 @@ describe(name, () => {
       );
 
     it('should be unchecked by default', () => {
-      const cb = mountStatelessCheckbox({ isChecked: false });
+      const cb = mountCheckbox({ defaultChecked: false });
       expect(cb.find('input[checked]').length === 1).toBe(true);
     });
     it('should call onchange on change', () => {
       const myMock = jest.fn();
-      const cb = mountStatelessCheckbox({ isChecked: false, onChange: myMock });
-      cb.find(HiddenCheckbox).simulate('change', { target: { checked: true } });
-      expect(cb.find('CheckboxStateless').prop('isChecked')).toBe(false);
+      const cb = mountCheckbox({ isChecked: false, onChange: myMock });
+      cb.find(Checkbox).simulate('change', { target: { checked: true } });
+      expect(cb.find('Checkbox').prop('isChecked')).toBe(false);
       expect(myMock.mock.calls.length).toBe(1);
     });
     it('should show indeterminate icon when indeterminate', () => {
-      const cb = mountStatelessCheckbox({
-        isChecked: false,
-        isIndeterminate: true,
-      });
+      const cb = mount(
+        <Checkbox
+          onChange={() => {}}
+          name="stub"
+          value="stub value"
+          isIndeterminate
+          isChecked={false}
+        />,
+      );
+
       expect(cb.find(CheckboxIcon)).toHaveLength(0);
       expect(cb.find(CheckboxIndeterminateIcon)).toHaveLength(1);
     });
     it('should initially set the indeterminate state on the hidden checkbox', () => {
-      const cb = mountStatelessCheckbox({
-        isChecked: false,
-        isIndeterminate: true,
-      });
-      const element = cb.find('CheckboxStateless').instance().checkbox;
+      const cb = mount(
+        <Checkbox
+          onChange={() => {}}
+          name="stub"
+          value="stub value"
+          isIndeterminate
+          isChecked={false}
+        />,
+      );
+      const element = cb.find('Checkbox').instance().checkbox;
       expect(element.indeterminate).toBe(true);
     });
     it('should set the indeterminate state on the hidden checkbox on update', () => {
-      const cb = mountStatelessCheckbox({
+      const cb = mountCheckbox({
         isChecked: false,
         isIndeterminate: false,
       });
 
-      const element = cb.find('CheckboxStateless').instance().checkbox;
+      const element = cb.find('Checkbox').instance().checkbox;
       expect(element.indeterminate).toBe(false);
 
       cb.setProps({ isIndeterminate: true });
       expect(element.indeterminate).toBe(true);
     });
-    it('should have ref to CheckboxStateless class', () => {
-      const spy = jest.fn();
-      mount(
-        <div>
-          <CheckboxStateless
-            ref={spy}
-            label=""
-            isChecked
-            onChange={() => {}}
-            name="stub"
-            value="stub value"
-          />
-        </div>,
-      );
-      expect(spy).toHaveBeenCalled();
-      const [instance] = spy.mock.calls[0];
-      expect(instance).toBeInstanceOf(CheckboxStateless);
-    });
   });
-  describe('<Checkbox />', () => {
-    it('should render initiallyChecked', () => {
+  describe('<Checkbox defaultChecked/>', () => {
+    it('should render defaultChecked', () => {
       const cb = mount(
-        <Checkbox label="" name="stub" value="stub value" initiallyChecked />,
+        <Checkbox label="" name="stub" value="stub value" defaultChecked />,
       );
-      expect(cb.find('CheckboxStateless').prop('isChecked')).toBe(true);
+      const element = cb.find('Checkbox').instance().checkbox;
+      expect(element.checked).toBe(true);
     });
-    it('should render initiallyChecked={false}', () => {
+    it('should render defaultChecked={undefined}', () => {
       const cb = mount(<Checkbox label="" name="stub" value="stub value" />);
-      expect(cb.find('CheckboxStateless').prop('isChecked')).toBe(false);
-    });
-  });
-  describe('<CheckboxGroup />', () => {
-    it('sanity check for CheckboxGroup', () => {
-      const cb = mount(
-        <CheckboxGroup>
-          <Checkbox label="" name="stub" value="stub value" />
-          <Checkbox label="" name="stub2" value="stub value2" />
-          <Checkbox label="" name="stub3" value="stub value3" />
-          <Checkbox label="" name="stub4" value="stub value4" />
-        </CheckboxGroup>,
-      );
-      expect(cb.find(Checkbox).length).toBe(4);
+      const element = cb.find('Checkbox').instance().checkbox;
+      expect(element.checked).toBe(false);
     });
   });
 });
 
-describe('CheckboxStatelessWithAnalytics', () => {
+describe('CheckboxWithAnalytics', () => {
   beforeEach(() => {
     jest.spyOn(global.console, 'warn');
     jest.spyOn(global.console, 'error');
@@ -123,7 +99,7 @@ describe('CheckboxStatelessWithAnalytics', () => {
 
   it('should mount without errors', () => {
     mount(
-      <CheckboxStatelessWithAnalytics
+      <Checkbox
         label=""
         isChecked
         onChange={() => {}}