+++ packages/editor/editor-core/example-helpers/mock-collab-provider.ts
@@ -84,9 +84,9 @@ export class MockCollabEditProvider implements CollabEditProvider {
   protected getState = () => {};
   protected createStep = (json: object) => {};
   protected sid;
-  protected eventBus: EventEmitter;
+  protected eventBus: any;
 
-  constructor(eventBus: EventEmitter, sid?: string) {
+  constructor(eventBus: any, sid?: string) {
     // If there's no sid then it's single user, being used for test
     if (sid) {
       this.sid = sid;