+++ packages/editor/editor-core/src/plugins/tasks-and-decisions/ui/styles.ts
@@ -2,10 +2,12 @@
 // prettier-ignore
 import { css, Styles, StyledComponentClass } from 'styled-components';
 import { fontSize } from '@atlaskit/theme';
+import { akEditorElementMinWidth } from '@atlaskit/editor-common';
 
 export const tasksAndDecisionsStyles = css`
   .ProseMirror .taskItemView-content-wrap,
   .ProseMirror .decisionItemView-content-wrap {
     font-size: ${fontSize}px;
+    min-width: ${akEditorElementMinWidth}px;
   }
 `;