+++ packages/media/media-card/src/root/card/index.tsx
@@ -62,9 +62,9 @@ export class Card extends Component<CardProps, CardState> {
   };
 
   componentDidMount() {
-    const { identifier, context, dimensions } = this.props;
+    const { identifier, context } = this.props;
 
-    this.subscribe(identifier, context, dimensions);
+    this.subscribe(identifier, context);
   }
 
   componentWillReceiveProps(nextProps: CardProps) {
@@ -78,14 +78,22 @@ export class Card extends Component<CardProps, CardState> {
       identifier: nextIdenfifier,
       dimensions: nextDimensions,
     } = nextProps;
+
     if (
       currentContext !== nextContext ||
       !deepEqual(currentIdentifier, nextIdenfifier) ||
-      isBigger(currentDimensions, nextDimensions)
+      this.shouldRefetchImage(currentDimensions, nextDimensions)
     ) {
-      this.subscribe(nextIdenfifier, nextContext, nextDimensions);
+      this.subscribe(nextIdenfifier, nextContext);
     }
   }
+
+  shouldRefetchImage = (current?: CardDimensions, next?: CardDimensions) => {
+    if (!current || !next) {
+      return false;
+    }
+    return isBigger(current, next);
+  };
 
   componentWillUnmount() {
     this.unsubscribe();
@@ -111,11 +119,7 @@ export class Card extends Component<CardProps, CardState> {
     }
   };
 
-  async subscribe(
-    identifier: Identifier,
-    context: Context,
-    dimensions?: CardDimensions,
-  ) {
+  async subscribe(identifier: Identifier, context: Context) {
     const { isCardVisible } = this.state;
     if (!isCardVisible) {
       return;
@@ -178,7 +182,7 @@ export class Card extends Component<CardProps, CardState> {
               break;
             case 'processed':
               if (metadata.mediaType && isPreviewableType(metadata.mediaType)) {
-                const { appearance, resizeMode } = this.props;
+                const { appearance, dimensions, resizeMode } = this.props;
                 const options = {
                   appearance,
                   dimensions,
@@ -229,9 +233,9 @@ export class Card extends Component<CardProps, CardState> {
 
   // This method is called when card fails and user press 'Retry'
   private onRetry = () => {
-    const { identifier, context, dimensions } = this.props;
+    const { identifier, context } = this.props;
 
-    this.subscribe(identifier, context, dimensions);
+    this.subscribe(identifier, context);
   };
 
   private isFile(identifier: Identifier): identifier is FileIdentifier {
@@ -326,8 +330,8 @@ export class Card extends Component<CardProps, CardState> {
 
   onCardInViewport = () => {
     this.setState({ isCardVisible: true }, () => {
-      const { identifier, context, dimensions } = this.props;
-      this.subscribe(identifier, context, dimensions);
+      const { identifier, context } = this.props;
+      this.subscribe(identifier, context);
     });
   };
 }