+++ packages/search/global-search/src/components/home/SearchResultsState.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ResultItemGroup } from '@atlaskit/quick-search';
-import ConfluenceIcon from '@atlaskit/icon/glyph/confluence';
+import { ConfluenceIcon } from '@atlaskit/logo';
 import PeopleIcon from '@atlaskit/icon/glyph/people';
 import { Result } from '../../model/Result';
 import { take, isEmpty } from '../SearchResultsUtil';