+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/global-navigation
 
+## 4.4.2
+- [patch] fixing HelpIcon active on children :hover [ae95dad](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ae95dad)
+
 ## 4.4.1
 - [patch] Do not cache CustomComponent in GlobalItem [372795f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/372795f)